Revert "Implementation of user auth endpoints "
