+++ .gitignore
@@ -3,5 +3,4 @@
 .1024nix
 .1024feature*
 .nfs*
-.env
 node_modules
