+++ config/config.go
@@ -1,0 +1,7 @@
+package config
+
+const MYSQL_HOST = "172.16.32.92"
+const MYSQL_PASSWORD = "PtoYbtCt"
+const MYSQL_PORT = "50122"
+const MYSQL_USERNAME = "root"
+const MYSQL_NAME = "tiktok_backend"
