--- controllers/likeController.go
@@ -1,1 +1,0 @@
-package controllers
