+++ dao/video.go
@@ -1,0 +1,1 @@
+package dao
