+++ main.go
@@ -1,24 +1,20 @@
 package main
 
 import (
-	"fmt"
-	"github.com/gin-gonic/gin"
-
-	"main/services/database"
+  "fmt"
+  "github.com/gin-gonic/gin"
+  "main/dao"
 )
 
-func init() {
-	// Initiate project dependencies
-	services.InitDB()
-	services.SyncDB()
-}
-
 func main() {
-	r := gin.Default()
+  // Initiate project dependencies
+  dao.InitDB()
+  
+  r := gin.Default()
 
 	initRouter(r)
 
 	r.Run(":8080")
 
-	fmt.Println("Application running on port 8080")
+  fmt.Println("Application running on port 8080")
 }
