--- models/userModel.go
@@ -1,9 +1,0 @@
-package models
-
-import "gorm.io/gorm"
-
-type User struct {
-  gorm.Model
-  Username string `gorm:"unique"`
-  Password string
-}
